vfs: complete rewrite syscalls haven't been updated yet

+++ kernel/fs.c
@@ -2,11 +2,16 @@
 #include <stddef.h>
 #include <stdint.h>
 
+#include <list.h>
 #include <console.h>
 #include <fs.h>
 #include <heap.h>
+#include <string.h>
 
-fs_node_t *fs_root;
+/**
+root node of the mount graph
+*/
+fs_mount_t *fs_root_mount = NULL;
 
 fs_node_t *fs_alloc_node() {
 	fs_node_t *node = kcalloc(1, sizeof(fs_node_t));
@@ -19,7 +24,7 @@ fs_node_t *fs_alloc_node() {
 }
 
 void fs_free_node(fs_node_t **node) {
-	assert(*node != fs_root);
+	assert(*node != fs_root_mount->node);
 	if ((*node)->close) {
 		(*node)->close(*node);
 	}
@@ -33,7 +38,7 @@ void fs_open(fs_node_t *node, unsigned int flags) {
 	}
 
 	if (node->__refcount >= 0) {
-		// lock
+		// TODO: lock
 		node->__refcount++;
 	}
 
@@ -44,7 +49,7 @@ void fs_open(fs_node_t *node, unsigned int flags) {
 
 void fs_close(fs_node_t *node) {
 	assert(node != NULL); // crash on double-free
-	assert(node != fs_root);
+	assert(node != fs_root_mount->node);
 
 	if (node->__refcount == -1) {
 		// special case ( TODO: check for FS_NODE_MOUNTPOINT instead
@@ -55,34 +60,23 @@ void fs_close(fs_node_t *node) {
 	node->__refcount--;
 
 	if (node->__refcount == 0) {
-		printf("node refcount = 0\nfs_free_node: %p\n", node);
 		fs_free_node(&node);
 	}
 	return;
 }
 
-uint32_t fs_read(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer) {
-	if (!node) {
-		return -1;
+uint32_t fs_read(fs_node_t *node, uint32_t offset, uint32_t size, void *buffer) {
+	if ((node != NULL) && (node->read != NULL)) {
+		return node->read(node, offset, size, buffer);
 	}
-
-	if (node->read) {
-		return (uint32_t)node->read(node, offset, size, buffer);
-	} else {
-		return -1;
-	}
+	return -1;
 }
 
-uint32_t fs_write(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer) {
-	if (!node) {
-		return -1;
+uint32_t fs_write(fs_node_t *node, uint32_t offset, uint32_t size, void *buffer) {
+	if ((node != NULL) && (node->write != NULL)) {
+		return node->write(node, offset, size, buffer);
 	}
-
-	if (node->write) {
-		return (uint32_t)node->write(node, offset, size, buffer);
-	} else {
-		return -1;
-	}
+	return -1;
 }
 
 struct dirent *fs_readdir(struct fs_node *node, uint32_t i) {
@@ -92,9 +86,9 @@ struct dirent *fs_readdir(struct fs_node *node, uint32_t i) {
 
 	if ((node->flags & FS_NODE_DIRECTORY) && (node->readdir != NULL)) {
 		return node->readdir(node, i);
-	} else {
-		return NULL;
 	}
+
+	return NULL;
 }
 
 struct fs_node *fs_finddir(struct fs_node *node, char *name) {
@@ -121,31 +115,42 @@ void fs_create (fs_node_t *node, char *name, uint16_t permissions) {
 	return;
 }
 
-void fs_unlink (fs_node_t *node, char *name) {
-	if (!node) {
-		return;
+int fs_unlink (fs_node_t *node, char *name) {
+	if (node == NULL) {
+		return -1;
 	}
 
-	if (node->unlink) {
+	if ((node->flags & FS_NODE_DIRECTORY) && (node->unlink != NULL)) {
 		return node->unlink(node, name);
-	} else {
-		return;
 	}
+
+	return -1;
 }
 
 void fs_mkdir(fs_node_t *node, char *name, uint16_t permission) {
 	if (node == NULL) {
 		return;
 	}
-
-	if (node->mkdir) {
+	if ((node->flags & FS_NODE_DIRECTORY) && (node->mkdir != NULL)) {
 		return node->mkdir(node, name, permission);
-	} else {
-		return;
 	}
+
+	return;
 }
 
 void fs_mount_root(fs_node_t *node) {
-	fs_root = node;
+	// TODO: implement root remount
+	assert(fs_root_mount == NULL);
+	fs_root_mount = kcalloc(1, sizeof(fs_mount_t));
+	fs_root_mount->path = "/";
+	fs_root_mount->node = node;
+	fs_root_mount->mounts = NULL;
 }
 
+fs_node_t *kopen(char *path, unsigned int flags) {
+	if (node == NULL) {
+		return NULL;
+	}
+	fs_open(node, flags);
+	return node;
+}
